+++ CHANGES.md
@@ -1,6 +1,7 @@
 ## Changes in 0.1.2 (in development)
 
 * Improve handling of environment file specification (#63)
+* Stop running container on SIGINT (#62)
 
 ## Changes in 0.1.1